// src/enums/Modals.ts
export enum ModalsEnum {
    USERNAME_INPUT = 'username-input',
    GITHUB_USERNAME_INPUT_LABEL = 'GitHub Username *',
    GITHUB_USERNAME_INPUT_LABEL_DEFAULT = 'enter the GitHub username here',
    GITHUB_USERNAME_INPUT = 'github-username',
    GITHUB_USERNAME_BLOCK = 'github-username-block',
    SAVE_USERNAME_SUBMIT_BUTTON_LABEL = 'Save Username',
    SAVE_USERNAME = 'save-username',
    USERNAME_INPUT_MODAL_NAME = 'Save GitHub Username',
    
    PENDING_USER_MAPPING_APPROVALS = 'pending-user-mapping-approvals',
    PENDING_USER_MAPPING_APPROVALS_MODAL_NAME = 'Pending User GitHub Mapping Approvals'
}